Work with save model method (not finished)
