+++ Modules/Administration/Models/AclUsersModel.php
@@ -10,10 +10,12 @@ namespace Asmoria\Modules\Administration\Models;
 
 use Asmoria\Core\Model;
 use Asmoria\Modules\Profiler\ProfilerController;
+use Asmoria\Modules\Handler\HandlerController;
 use Asmoria\Modules\Administration\Models\AclRolesModel as Roles;
 
 
-class AclUsersModel extends Model{
+class AclUsersModel extends Model
+{
 
     static $_instance;
     public $idField = "usr_id";
@@ -22,8 +24,7 @@ class AclUsersModel extends Model{
 
     public function __construct()
     {
-//        parent::__construct();
-//        $this->profile = ProfilerController::getInstance();
+        parent::__construct();
     }
 
 
@@ -35,15 +36,15 @@ class AclUsersModel extends Model{
     public function isAdmin($id)
     {
         $role = $this->getById($id);
-        if(empty($role))
-            throw new \Exception("Cannot find user role");
+        if (empty($role))
+            throw new HandlerController(new \Exception("Cannot find user role"));
         $role = Roles::getInstance()->getById($role->role_id);
-        return $role->type === ADMIN_ROLE ? TRUE : FALSE;
+        return intval($role->id) === ADMIN_ROLE ? TRUE : FALSE;
     }
 
-    public function test()
+    public function setRole($user, $role)
     {
-       return $this->getById(1);
+
     }
 
     public static function getInstance()
