+++ Modules/Profiler/ProfilerController.php
@@ -9,23 +9,23 @@
 namespace Asmoria\Modules\Profiler;
 
 use Asmoria\Core\Configuration;
-use Asmoria\Core\Model;
+use Asmoria\Core\Controller;
+use Asmoria\Modules\Handler\HandlerController;
 use Asmoria\Modules\Profiler\Models\ProfileModel;
 use Asmoria\Modules\Handler\HandlerController as Handler;
 use Asmoria\Modules\Administration\Models\AclUsersModel as UsersRole;
 
 
-class ProfilerController
+class ProfilerController extends Controller
 {
     static $_instance;
-    private $db;
     public $profile;
     public $isAdmin;
 
 
     public function __construct()
     {
-        $this->db = Configuration::getInstance();
+        parent::__construct();
         $this->authorize();
         if(!empty($_SESSION['u_id']))
         $this->isAdmin = UsersRole::getInstance()->isAdmin($_SESSION['u_id']);
@@ -44,12 +44,12 @@ class ProfilerController
             return true;
         }
         $result = "";
+        if(isset($_POST['submit']))
         if (!empty($_POST['a_email']) && !empty($_POST['a_pass'])) {
-            $db = Configuration::getInstance();
             $mail = $_POST['a_email'];
             $pass = $this->db->encodePass($_POST['a_pass']);
             $model = new ProfileModel();
-            $sql_ = $db->connection->prepare("SELECT * FROM ".$model->table." WHERE `mail` = :mail AND `pass` = :pass");
+            $sql_ = $this->db->connection->prepare("SELECT * FROM ".$model->table." WHERE `mail` = :mail AND `pass` = :pass");
             $sql_->bindValue(':mail', $mail, \PDO::PARAM_STR);
             $sql_->bindValue(':pass', $pass, \PDO::PARAM_STR);
             $sql_->setFetchMode(\PDO::FETCH_CLASS, "Asmoria\\Modules\\Profiler\\Models\\ProfileModel");
@@ -60,7 +60,7 @@ class ProfilerController
                 $_SESSION['u_id'] = $check->id;
                 $_SESSION['u_mail'] = $check->mail;
                 $result2['status'] = 'ok';
-                $result2['content'] = $db->getLoginBar();
+                $result2['content'] = $this->db->getLoginBar();
                 echo json_encode($result2);exit;
             }
             else{
@@ -70,20 +70,102 @@ class ProfilerController
             }
         }
         else {
-            //Handler::getInstance(new \Exception("Register or authorize first"), true)->getError();
+            throw new HandlerController(new \Exception("Fill all fields first"), true);
         }
     }
 
-    public function actionIndex()
+
+    public function actionRegister()
     {
-        echo $this->db->getHeader();
-        echo "Index Here";
-        echo $this->db->getFooter();
+        $result = "";
+        if (!empty($_POST['mail']) && !empty($_POST['pass'])) {
+            $email = $_POST['mail'];
+            $pass = $_POST['pass'];
+            $conf_pass = $_POST['conf_pass'];
+
+            $passlen = strlen($pass);
+            $emaillen = strlen($email);
+            $max = 255;
+            $minpass = 6;
+            $minemail = 3;
+
+            if($passlen < $minpass){
+                $errors[] = "pass must be at least $minpass characters";
+            } elseif($passlen > $max){
+                $errors[] = "pass must be less than $max characters";
+            }
+
+            if($emaillen < $minemail){
+                $errors[] = "email must be at least $minemail characters";
+            } elseif($emaillen > $max){
+                $errors[] = "email must be less than $max characters";
+            }
+
+            if($pass != $conf_pass){
+                $errors[] = "your passwords do not match";
+            }
+
+            if(empty($pass)){
+                $errors[] = "pass is required";
+            }
+
+            if(empty($email)){
+                $errors[] = "email cannot be left empty";
+            }
+
+            if(!filter_var($email, FILTER_VALIDATE_EMAIL)){
+                $errors[] = "invalid email";
+            }
+            if (!empty($errors)) {
+//                echo json_encode($errors);
+                $result['errors'] = "<ul>";
+                foreach ($errors as $error) {
+                    $result['errors'] .= "<li>$error</li>";
+                }
+                $result['errors'] .= "</ul>";
+                echo json_encode($result);
+                exit;
+            }
+
+
+            $pass = $this->db->encodePass($pass);
+
+            $sql_ = $this->db->connection->prepare("INSERT INTO `profiles` (`mail`, `pass`)VALUES(:mail, :pass)");
+            $sql_->bindValue(':mail', $email, \PDO::PARAM_STR);
+            $sql_->bindValue(':pass', $pass, \PDO::PARAM_STR);
+            $result['status'] = $sql_->execute();
+            $sql_ = $this->db->connection->prepare("SELECT * FROM `profiles` WHERE `mail` = :mail AND `pass` = :pass");
+            $sql_->bindValue(':mail', $email, \PDO::PARAM_STR);
+            $sql_->bindValue(':pass', $pass, \PDO::PARAM_STR);
+            $result['status'] = $sql_->execute();
+            $result['data'] = $sql_->fetch(\PDO::FETCH_ASSOC);
+        }
+
+        if ($result) {
+
+            if(!isset($_SESSION))
+                session_start();
+            $_SESSION['u_id'] = $result['data']['id'];
+            $_SESSION['u_mail'] = $result['data']['mail'];
+            $result['loginBar'] = $this->db->getLoginBar();
+            $result['content'] = "Welcome to aboard";
+            echo json_encode($result);
+            exit;
+        } else die('Error');
+
     }
 
-    public static function test()
+
+
+    public function actionLogout(){
+        session_destroy();
+        header('Location: '.ROOT_URL);
+    }
+
+
+    public function actionIndex()
     {
-        echo "Hi Jack";exit;
+        $this->view->render('index');
     }
 
     public static function getInstance()
