+++ style/css/custom.css
@@ -25,7 +25,7 @@ footer, .push {
     left: 0;
     margin: 0 auto;
     background-color: #000000;
-    /*padding: 10px 0 10px 100px;*/
+    padding: 10px 0 10px 100px;
     border-top: 1px solid #eee;
 }
